Repository modified

+++ app/src/main/java/com/github/brugapp/brug/data/ItemsRepository.kt
@@ -2,6 +2,7 @@ package com.github.brugapp.brug.data
 
 import android.util.Log
 import com.github.brugapp.brug.model.MyItem
+import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.QueryDocumentSnapshot
 import kotlinx.coroutines.tasks.await
@@ -24,8 +25,10 @@ object ItemsRepository {
     suspend fun addItemToUser(
         item: MyItem,
         uid: String,
-        firestore: FirebaseFirestore
+        firestore: FirebaseFirestore,
+        isCached: Boolean = false
     ): FirebaseResponse {
+        Log.d("ItemsRepository", "addItemToUser: $item, $uid")
         val response = FirebaseResponse()
 
         try {
@@ -35,18 +38,38 @@ object ItemsRepository {
                 return response
             }
 
-            userRef.collection(ITEMS_DB).add(
-                mapOf(
-                    "item_name" to item.itemName,
-                    "item_type" to item.itemTypeID,
-                    "item_description" to item.itemDesc,
-                    "is_lost" to item.isLost()
-                )
-            ).await()
+            if(isCached) {
+                //control if items is already in the databaase
+                //TODO: control if the item in the database is the same
+                val itemRef = userRef.collection(ITEMS_DB).document(item.getItemID())
+                if (itemRef.get().await().exists()) {
+                    itemRef.update("is_deleted", false).await()
+                    response.onSuccess = true
+                }else{
+                    response.onSuccess = false
+                }
 
-            response.onSuccess = true
+            }
+            else{
+                Log.d("ItemRepository","Is not cached")
+                val id = userRef.collection(ITEMS_DB).add(
+                    mapOf(
+                        "item_name" to item.itemName,
+                        "item_type" to item.itemTypeID,
+                        "item_description" to item.itemDesc,
+                        "is_lost" to item.isLost(),
+                        "is_deleted" to item.getIsDeleted(),
+                        "deleted_date" to item.getDeletedWhen()
+                    )
+                ).await().id
+                Log.d("ITEM_ID",id)
+                item.setItemID(id)
+
+                response.onSuccess = true
+            }
         } catch (e: Exception) {
             response.onError = e
+            Log.d("ItemRepository","We have a problem amirite ${e}")
         }
 
         return response
@@ -84,7 +107,9 @@ object ItemsRepository {
                     "item_name" to item.itemName,
                     "item_type" to item.itemTypeID,
                     "item_description" to item.itemDesc,
-                    "is_lost" to item.isLost()
+                    "is_lost" to item.isLost(),
+                    "is_deleted" to item.getIsDeleted(),
+                    "deleted_date" to item.getDeletedWhen()
                 )
             ).await()
             response.onSuccess = true
@@ -106,7 +131,8 @@ object ItemsRepository {
     suspend fun deleteItemFromUser(
         itemID: String,
         uid: String,
-        firestore: FirebaseFirestore
+        firestore: FirebaseFirestore,
+        isCached: Boolean = false
     ): FirebaseResponse {
         val response = FirebaseResponse()
 
@@ -123,8 +149,14 @@ object ItemsRepository {
                 response.onError = Exception("Item doesn't exist")
                 return response
             }
+            if(!isCached) {
+                itemRef.delete().await()
+            }else{
+                //Do not forget to modify the item's is_deleted field
+                // here we only modify the is_deleted field in the database
+                //itemRef.update("item_description", "Pourquoi ca ne marche pas").await()
 
-            itemRef.delete().await()
+            }
 
             response.onSuccess = true
         } catch (e: Exception) {
@@ -155,7 +187,9 @@ object ItemsRepository {
                     "item_name" to item.itemName,
                     "item_type" to item.itemTypeID,
                     "item_description" to item.itemDesc,
-                    "is_lost" to item.isLost()
+                    "is_lost" to item.isLost(),
+                    "is_deleted" to item.getIsDeleted(),
+                    "deleted_date" to item.getDeletedWhen()
                 )
             ).await()
 
@@ -198,6 +232,19 @@ object ItemsRepository {
      * @return List<MyItem> containing all the user's items, or a null value in case of error.
      */
     suspend fun getUserItemsFromUID(uid: String, firestore: FirebaseFirestore): List<MyItem>? {
+        return getListOption(firestore, uid) { !it.getIsDeleted() }
+    }
+
+    suspend fun getUserDeletedItemsFromUid(uid:String, firestore: FirebaseFirestore): List<MyItem> ? {
+        return getListOption(firestore,uid) {it.getIsDeleted()}
+    }
+
+
+    private suspend fun getListOption(
+        firestore: FirebaseFirestore,
+        uid: String,
+        option: (MyItem) -> Boolean
+    ): List<MyItem>? {
         return try {
             val userRef = firestore.collection(USERS_DB).document(uid)
             if (!userRef.get().await().exists()) {
@@ -205,13 +252,27 @@ object ItemsRepository {
                 return null
             }
 
+            deleteOldDeletedItems(userRef)
+
             userRef.collection(ITEMS_DB).get().await().mapNotNull { item ->
                 getItemFromDoc(item)
-            }
+            }.filter(option)
 
         } catch (e: Exception) {
             Log.e("FIREBASE ERROR", e.message.toString())
             null
+        }
+    }
+
+
+    private suspend fun deleteOldDeletedItems(userRef: DocumentReference) {
+        userRef.collection(ITEMS_DB).get().await().mapNotNull { item ->
+            val myItem = getItemFromDoc(item)
+            if (myItem != null) {
+                if (myItem.isTooOld()) {
+                    userRef.collection(ITEMS_DB).document(item.id).delete().await()
+                }
+            }
         }
     }
 
@@ -222,8 +283,11 @@ object ItemsRepository {
                 || !itemDoc.contains("item_type")
                 || !itemDoc.contains("item_description")
                 || !itemDoc.contains("is_lost")
+                || !itemDoc.contains("is_deleted")
+                || !itemDoc.contains("deleted_date")
             ) {
                 Log.e("FIREBASE ERROR", "Invalid Item Format")
+                Log.e("ITEM IS","${itemDoc.data}" )
                 return null
             }
 
@@ -231,12 +295,14 @@ object ItemsRepository {
                 itemDoc["item_name"] as String,
                 (itemDoc["item_type"] as Long).toInt(),
                 itemDoc["item_description"] as String,
-                itemDoc["is_lost"] as Boolean
+                itemDoc["is_lost"] as Boolean,
             )
             item.setItemID(itemDoc.id)
+            item.setDeleted(itemDoc["is_deleted"] as Boolean)
+            item.setDeletedWhen(itemDoc["deleted_date"] as String)
             return item
         } catch (e: Exception) {
-            Log.e("FIREBASE ERROR", e.message.toString())
+            Log.e("FIREBASE ERROR TRANSFORMATION", e.message.toString())
             return null
         }
     }
